requirements.txt
requests>=2.31
pyyaml>=6.0
Pillow>=10.0
